tourapp/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
Deprecated==1.2.13
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
